Move transformations to separate subpackage
